refactor: Move fee-to-dead mechanism to AcctCtrler
