+++ ctrlers/supply/ctrler_block.go
@@ -4,7 +4,6 @@ import (
 	ctrlertypes "github.com/beatoz/beatoz-go/ctrlers/types"
 	v1 "github.com/beatoz/beatoz-go/ledger/v1"
 	"github.com/beatoz/beatoz-go/types/xerrors"
-	"github.com/holiman/uint256"
 	abcitypes "github.com/tendermint/tendermint/abci/types"
 	"time"
 )
@@ -27,44 +26,6 @@ func (ctrler *SupplyCtrler) EndBlock(bctx *ctrlertypes.BlockContext) ([]abcitype
 	defer ctrler.mtx.Unlock()
 
 	var evts []abcitypes.Event
-
-	//
-	// Process txs fee: Burn and Reward
-	header := bctx.BlockInfo().Header
-	sumFee := bctx.SumFee() // it's value is 0 at BeginBlock.
-	if header.GetProposerAddress() != nil && sumFee.Sign() > 0 {
-
-		//
-		// Reward the GovParams.TxFeeRewardRate % of txs fee to the proposer of this block.
-		rwdAmt := new(uint256.Int).Mul(sumFee, uint256.NewInt(uint64(bctx.GovHandler.TxFeeRewardRate())))
-		rwdAmt = new(uint256.Int).Div(rwdAmt, uint256.NewInt(100))
-		if xerr := bctx.AcctHandler.AddBalance(header.GetProposerAddress(), rwdAmt, true); xerr != nil {
-			return nil, xerr
-		}
-
-		//
-		// Auto burn(dead): transfer the remaining fee to DEAD Address
-		//
-		// this is not burning.
-		// it is just to transfer to the zero address.
-		//// In ctrler.burn, ctrler.lastTotalSupply is changed.
-		//if xerr := ctrler.burn(bctx.Height(), burnAmt); xerr != nil {
-		//	return nil, xerr
-		//}
-
-		deadAmt := new(uint256.Int).Sub(sumFee, rwdAmt)
-		if xerr := bctx.AcctHandler.AddBalance(bctx.GovHandler.DeadAddress(), deadAmt, true); xerr != nil {
-			return nil, xerr
-		}
-		evts = append(evts, abcitypes.Event{
-			Type: "supply.txfee",
-			Attributes: []abcitypes.EventAttribute{
-				{Key: []byte("dead"), Value: []byte(deadAmt.Dec()), Index: false},
-				{Key: []byte("reward"), Value: []byte(rwdAmt.Dec()), Index: false},
-			},
-		})
-		ctrler.logger.Debug("txs's fee is processed", "total.fee", sumFee.Dec(), "reward", rwdAmt.Dec(), "dead", deadAmt.Dec())
-	}
 
 	//
 	// Wait to finish minting...
